so Year and Time together

+++ manipulate-count.r
@@ -14,7 +14,7 @@ data$Dayte <- dayte(data$Date)
 data$Replicate <- factor(data$Replicate)
 data$Year <- factor(data$Year)
 
-data$Time <- as.numeric(as.character(data$Year))
+data$Time <- as.integer(as.character(data$Year))
 
 save_rdata(data)
